Add database seeder files

+++ database/migrations/2019_08_11_153930_create_patients_table.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * File for creating patients table
- * php version 7.2.19
+ * PHP version 7.2.19
  *
  * @category Patients_Migration
  * @package  Migrations
@@ -16,7 +16,7 @@ use Illuminate\Database\Migrations\Migration;
 
 /**
  * Class for creating patients table
- * php version 7.2.19
+ * PHP version 7.2.19
  *
  * You may add fields to patients table here
  *
@@ -40,7 +40,7 @@ class CreatePatientsTable extends Migration
             'patients',
             function (Blueprint $table) {
                 $table->bigIncrements('id');
-                $table->integer('phn');
+                $table->bigInteger('phn');
                 $table->string('name')->nullable();
                 $table->string('nic')->nullable();
                 $table->timestamps();
